Windows fixes, plugin version fixes, warning fixes, lata fixes.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@16821 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/MetaData/avtScalarMetaData.C
@@ -745,7 +745,7 @@ avtScalarMetaData::AddEnumNameValue(std::string name, double val)
     enumRanges.push_back(val);
     enumRanges.push_back(val);
 
-    return enumNames.size() - 1;
+    return static_cast<int>(enumNames.size()) - 1;
 }
 
 // ****************************************************************************
@@ -915,7 +915,7 @@ avtScalarMetaData::ComboValFromDigits(const std::vector<std::vector<int> > &ptMa
     int row, col;
     *id = 0.0;
     std::list<int>::const_reverse_iterator it;
-    for (col = digits.size()-1, it = digits.rbegin(); col >= 0; col--, it++)
+    for (col = static_cast<int>(digits.size()-1), it = digits.rbegin(); col >= 0; col--, it++)
     {
         for (row = 0; row <= *it; row++)
             *id += ptMap[row][col];
